Refactor: restructure based on comments by @hoh in PR
